requirements.txt
pyserial>=3.5
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
streamlit>=1.28.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
